- Refactor call handling by replacing `CallAdapter` and `CallAudioController` with `CallOrchestrator` and `CallServiceAudioDelegate` - Implement `CallForegroundService` to handle incoming, outgoing, and ongoing call notifications as a foreground service - Introduce `CallUiEffects` and `CallUiEffectsHandler` to decouple UI and notification logic from the core call orchestration - Update `AndroidManifest.xml` with required `FOREGROUND_SERVICE` and `FOREGROUND_SERVICE_PHONE_CALL` permissions - Migrate call state management and lifecycle callbacks to `DefaultCallOrchestrator` - Refactor `CallService` to delegate call lifecycle events and audio controls to the orchestrator - Update `CallState` to use `Int` for duration and refine state mapping logic - Clean up `NotificationController` by moving display and cancellation logic to the foreground service handler

+++ app/src/main/java/com/talsk/amadz/ui/ongoingCall/CallScreen.kt
@@ -198,7 +198,7 @@ fun CallHeader(
         Spacer(modifier = Modifier.height(12.dp))
         if (uiState is CallState.Active) {
             Text(
-                text = secondsToReadableTime(uiState.duration.toInt()),
+                text = secondsToReadableTime(uiState.duration),
                 style = MaterialTheme.typography.bodySmall,
                 fontWeight = FontWeight.Bold
             )
